+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [101.1.14] - 2023-05-01
 ### Changed
 - code quality improvements
+- fix image helper for not existing files 
 
 ## [101.1.13] - 2023-04-21
 ### Fixed