+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Added
 
+## [Unreleased]
+### Changed
+- code quality improvements
+
 ## [101.1.15] - 2023-05-08
 ### Changed
 - code quality improvements