+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 ### Added
 
+## [101.1.16] - 2023-05-09
+### Fixed
+- fix for error "Email template '' is not defined" in magento 2.4.0
+### Changed
+- code quality improvements
+
 ## [101.1.15] - 2023-05-08
 ### Changed
 - code quality improvements